+++ src/generators/string.js
@@ -1,4 +1,5 @@
 import { LoremIpsum } from "lorem-ipsum";
+import { generateString } from "./util/utilities";
 
 const lorem = new LoremIpsum();
 
@@ -8,4 +9,25 @@ export function stringLoremSentence() {
 
 export function stringLoremParagraph() {
   return lorem.generateParagraphs(1);
+}
+
+export function stringAlphanumeric() {
+  return generateString(
+    nova.config.get("random.stringLength", "number") || 32,
+    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
+  );
+}
+
+export function stringAlphabetical() {
+  return generateString(
+    nova.config.get("random.stringLength", "number") || 32,
+    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
+  );
+}
+
+export function stringPassword() {
+  return generateString(
+    nova.config.get("random.stringLength", "number") || 32,
+    `ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!@£#$%^&*()_-+=[]{}\|;:'",.<>/?`
+  );
 }