+++ src/generators/util/utilities.js
@@ -3,3 +3,14 @@ export function generateNumber(min, max) {
   max = Math.floor(max);
   return Math.floor(Math.random() * (max - min + 1) + min);
 }
+
+export function generateString(
+  length,
+  charString = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
+) {
+  let str = "";
+  for (let i = 0; i < length; i++) {
+    str += charString.charAt(Math.floor(Math.random() * charString.length));
+  }
+  return str;
+}